+++ R/size.R
@@ -12,6 +12,12 @@
 #' frame (even data frame and matrix columns) have the same size.
 #' `vec_size_common(...)` returns the common size of multiple vectors.
 #'
+#' `list_sizes()` returns an integer vector containing the size of each element
+#' of a list. It is nearly equivalent to, but faster than,
+#' `map_int(x, vec_size)`, with the exception that `list_sizes()` will
+#' error on non-list inputs, as defined by [vec_is_list()]. `list_sizes()` is
+#' to `vec_size()` as [lengths()] is to [length()].
+#'
 #' @seealso [vec_slice()] for a variation of `[` compatible with `vec_size()`,
 #'   and [vec_recycle()] to recycle vectors to common length.
 #' @section Invariants:
@@ -68,6 +74,8 @@
 #' vec_size_common(1:10, 1:10)
 #' vec_size_common(1:10, 1)
 #' vec_size_common(integer(), 1)
+#'
+#' list_sizes(list("a", 1:5, letters))
 vec_size <- function(x) {
   .Call(vctrs_size, x)
 }
@@ -76,6 +84,12 @@ vec_size <- function(x) {
 #' @rdname vec_size
 vec_size_common <- function(..., .size = NULL, .absent = 0L) {
   .External2(vctrs_size_common, .size, .absent)
+}
+
+#' @rdname vec_size
+#' @export
+list_sizes <- function(x) {
+  .Call(vctrs_list_sizes, x)
 }
 
 #' @rdname vec_size